# internal/adapters/http/router.go
package http

import (
	"context"
	"log/slog"
	"net/http"
	"strings"
	"time"

	"github.com/google/uuid"

	"github.com/partyzanex/padmark/internal/adapters/http/ogenapi"
)

type contextKey uint8

const keyRequestID contextKey = 1

// NewRouter registers all routes and wraps them with middleware.
// The ogen-generated server handles JSON API routes (POST/PUT/DELETE /notes, GET /notes/{id} JSON,
// /healthz, /readyz). Manual handlers serve HTML pages, content-negotiated views, login, and static.
func NewRouter(handler *Handler, ogenHandler *OgenHandler, tokens []string) http.Handler {
	ogenSrv, err := ogenapi.NewServer(ogenHandler)
	if err != nil {
		panic("ogen server: " + err.Error())
	}

	mux := http.NewServeMux()

	// Ogen-handled JSON API routes
	mux.Handle("POST /notes", ogenSrv)
	mux.Handle("PUT /notes/{id}", ogenSrv)
	mux.Handle("DELETE /notes/{id}", ogenSrv)
	mux.Handle("GET /healthz", ogenSrv)
	mux.Handle("GET /readyz", ogenSrv)

	// Manual handlers: HTML pages + content-negotiated GET + login + static + api docs
	mux.HandleFunc("GET /login", handler.LoginPage)
	mux.HandleFunc("POST /login", loginHandler(makeTokenSet(tokens)))
	mux.HandleFunc("GET /api", handler.APIDocsPage)
	mux.HandleFunc("GET /api/openapi.yaml", APISpec)
	mux.HandleFunc("GET /", handler.IndexPage)
	mux.HandleFunc("GET /success", handler.SuccessPage)
	mux.Handle("GET /static/", StaticHandler)
	mux.HandleFunc("GET /notes/{id}", handler.GetNote)
	mux.HandleFunc("GET /edit/{id}", handler.EditPage)
	mux.HandleFunc("GET /{id}", handler.GetNote)

	stack := withRecovery(handler.log, mux)
	stack = newAuthMiddleware(tokens, stack)
	stack = withLogging(handler.log, stack)
	stack = withRequestID(stack)

	return stack
}

func makeTokenSet(tokens []string) map[string]struct{} {
	set := make(map[string]struct{}, len(tokens))

	for _, tok := range tokens {
		set[tok] = struct{}{}
	}

	return set
}

func withRequestID(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		id := uuid.NewString()
		w.Header().Set("X-Request-ID", id)
		next.ServeHTTP(w, r.WithContext(context.WithValue(r.Context(), keyRequestID, id)))
	})
}

type statusRecorder struct {
	http.ResponseWriter

	status  int
	written bool
}

func (sr *statusRecorder) WriteHeader(status int) {
	if !sr.written {
		sr.status = status
		sr.written = true
	}

	sr.ResponseWriter.WriteHeader(status)
}

func (sr *statusRecorder) Write(b []byte) (int, error) {
	if !sr.written {
		sr.status = http.StatusOK
		sr.written = true
	}

	return sr.ResponseWriter.Write(b) //nolint:wrapcheck // pass-through implementation of io.Writer
}

func withLogging(log *slog.Logger, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		rec := &statusRecorder{ResponseWriter: w, status: http.StatusOK}

		next.ServeHTTP(rec, r)

		level := slog.LevelInfo
		if strings.HasPrefix(r.URL.Path, "/static/") {
			level = slog.LevelDebug
		}

		log.Log(r.Context(), level, "http",
			"method", r.Method,
			"path", r.URL.Path,
			"status", rec.status,
			"duration", time.Since(start),
		)
	})
}

func withRecovery(log *slog.Logger, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()

		defer func() {
			if p := recover(); p != nil {
				log.ErrorContext(ctx, "panic recovered", "panic", p)
				http.Error(w, "internal server error", http.StatusInternalServerError)
			}
		}()

		next.ServeHTTP(w, r)
	})
}