# internal/adapters/http/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Padmark API
  description: JSON API for creating, storing, and managing Markdown notes.
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: notes
    description: CRUD operations on notes
  - name: health
    description: Liveness and readiness probes

paths:
  /notes:
    post:
      operationId: createNote
      summary: Create a new note
      tags: [notes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note created
          headers:
            Location:
              description: Canonical URL of the created note
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateNoteResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

  /notes/{id}:
    get:
      operationId: getNote
      summary: Get a note by ID
      description: Returns a note as JSON. Increments the view counter.
      tags: [notes]
      parameters:
        - $ref: "#/components/parameters/noteId"
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteResponse"
        "404":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

    put:
      operationId: updateNote
      summary: Update an existing note
      description: Requires the edit code that was returned when the note was created.
      tags: [notes]
      parameters:
        - $ref: "#/components/parameters/noteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

    delete:
      operationId: deleteNote
      summary: Delete a note
      description: Requires the edit code via X-Edit-Code header or edit_code query parameter.
      tags: [notes]
      parameters:
        - $ref: "#/components/parameters/noteId"
        - name: X-Edit-Code
          in: header
          schema:
            type: string
        - name: edit_code
          in: query
          schema:
            type: string
      responses:
        "204":
          description: Note deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

  /healthz:
    get:
      operationId: healthz
      summary: Liveness probe
      tags: [health]
      security: []
      responses:
        "200":
          description: Service is alive

  /readyz:
    get:
      operationId: readyz
      summary: Readiness probe
      tags: [health]
      security: []
      responses:
        "200":
          description: Service is ready
        "503":
          $ref: "#/components/responses/Error"

components:
  parameters:
    noteId:
      name: id
      in: path
      required: true
      schema:
        type: string

  schemas:
    CreateNoteRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        content_type:
          type: string
          enum: ["text/markdown", "text/plain"]
          default: "text/markdown"
        slug:
          type: string
        ttl:
          type: integer
          format: int64
        burn_after_reading:
          type: boolean
          default: false

    UpdateNoteRequest:
      type: object
      required: [title, content, edit_code]
      properties:
        title:
          type: string
        content:
          type: string
        content_type:
          type: string
          enum: ["text/markdown", "text/plain"]
        edit_code:
          type: string
        ttl:
          type: integer
          format: int64
        burn_after_reading:
          type: boolean

    NoteResponse:
      type: object
      required: [id, title, content, content_type, views, burn_after_reading, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        content_type:
          type: string
          enum: ["text/markdown", "text/plain"]
        views:
          type: integer
        burn_after_reading:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time

    CreateNoteResponse:
      type: object
      required: [id, title, content, content_type, views, burn_after_reading, created_at, updated_at, edit_code]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        content_type:
          type: string
          enum: ["text/markdown", "text/plain"]
        views:
          type: integer
        burn_after_reading:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        edit_code:
          type: string

    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"